test: Moved the extension mixin test into its own, separate file for easier testing in the future
